Add files via upload
I added an option for fitting columns with a round, 2D "Laplacian of Gaussian" function instead of the standard elliptical, rotatable, 2D Gaussian. 

To use this option, set the keyword arg for: AtomColumnLattice.fit_atom_columns(..., use_LoG_fitting=True). Default is False.

plot_fitting_residuals() method will automatically use the correct function (Gaussian or LoG) for calculating the residuals.

This MAY produce better results on dDPC images, but take care to check results. It SHOULD NOT be used on HAADF or iDPC images.

Note that no column shape information is yielded by this method (LoG function implemented is round) and integrated Gaussian intensity is NOT indicative of atomic number.