Fitting step timing and tweaks
Added timings for each step in the fitting prep routine. This is only temporary and will be removed in the future. The purpose is to identify slow sections for possible improvement.

Added a wrapper function for selecting a folder path using QFileDialog.

Added "fast_rotate_90deg" function to rotate images by 90 deg increments. It is faster than using functions that can rotate by arbitrary angles and therefore need to interpolate.

Advanced the version number to 2.1.3