Fixes and area masking for lattice generation
Fixes to UnitCell to plot correctly for single element lattices.

Addition of a band pass filter generator in utils.py and cleanup of get_phase_from_com.

Addition of mask creation methods to select image sub-areas for atom column fitting. Basic idea is to make a binary mask indicating the portion of the image in which the user wants to register a lattice and fit atom columns. When the reference lattice is generated, points inside the area are kept while those outside are discarded. The mask must be 1 where the lattice is to be kept, 0 elsewhere. The mask must be added to the AtomColumnLattice object as self.mask before define_reference_lattice() is called.