Back to old watershed, automatic scale determination
OpenCV watershed algorithm was removed in favor of sklearn's (which was perviously used). Reason is that OpenCV's does not function correctly for segmenting peaks in DPs or HRSTEM images. While it is slower, the results are better/correct when using sklearn's.

An automatic scale determination algorithm has been implemented. In most cases, this finds the correct sigma for filtering of images and DPs for peak location and watershed segmentation. The sigma values can still be specified, but the probe_fwhm attribute has been removed as it is not necessary. Now passing "auto" to the appropriate arguments uses the "select_scale()" method to find the right value.